+++ lib/extensions/reset_password/ecto/context.ex
@@ -7,9 +7,9 @@ defmodule PowResetPassword.Ecto.Context do
   def get_by_email(email, config), do: Operations.get_by([email: email], config)
 
   @spec update_password(Context.user(), map(), Config.t()) :: {:ok, Context.user()} | {:error, Context.changeset()}
-  def update_password(user, params, config) do
+  def update_password(%user_mod{} = user, params, config) do
     user
-    |> Schema.reset_password_changeset(params)
+    |> user_mod.reset_password_changeset(params)
     |> Context.do_update(config)
   end