+++ lib/extensions/reset_password/ecto/schema.ex
@@ -10,6 +10,18 @@ defmodule PowResetPassword.Ecto.Schema do
     Schema.require_schema_field!(module, :email, PowResetPassword)
   end
 
+  @doc false
+  @impl true
+  defmacro __using__(_config) do
+    quote do
+      def reset_password_changeset(changeset, attrs), do: pow_reset_password_changeset(changeset, attrs)
+
+      defdelegate pow_reset_password_changeset(changeset, attrs), to: unquote(__MODULE__), as: :reset_password_changeset
+
+      defoverridable reset_password_changeset: 2
+    end
+  end
+
   @spec reset_password_changeset(map(), map()) :: Changeset.t()
   def reset_password_changeset(%user_mod{} = user, params) do
     user