+++ lib/extensions/reset_password/reset_password.ex
@@ -6,6 +6,9 @@ defmodule PowResetPassword do
   def ecto_schema?(), do: true
 
   @impl true
+  def use_ecto_schema?(), do: true
+
+  @impl true
   def phoenix_messages?(), do: true
 
   @impl true